requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.13.2
anyio==4.4.0
arrow==1.3.0
ase==3.22.1
astroid==3.2.4
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
autopep8==2.0.4
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.40.0.post2
black==24.8.0
blessed==1.20.0
boto3==1.34.150
botocore==1.34.150
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cmake==3.26.4
colorama==0.4.5
colorlog==6.8.2
contextlib2==21.6.0
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1686733817565/work
croniter==1.3.15
cryptography==38.0.1
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dateutils==0.6.12
debugpy==1.5.1
decorator==5.1.1
deepdiff==7.0.1
dgl==2.0.0+cu117
dglgo==0.0.2
diffusers==0.17.1
dill==0.3.9
dnspython==2.6.1
docker-pycreds==0.4.0
docutils==0.20.1
dpcpp-cpp-rt==2023.0.0
editor==1.6.6
einops==0.7.0
emmet-core==0.68.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.88.0
fastjsonschema==2.16.2
filelock==3.12.2
flake8==7.1.1
fonttools==4.25.0
freetype-py @ file:///home/conda/feedstock_root/build_artifacts/freetype-py_1650983368720/work
frozenlist==1.4.1
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.1
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.22.0
imagesize==1.4.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==2.0.0
inquirer==3.3.0
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
ipykernel==6.15.2
ipython==8.4.0
ipywidgets==8.0.2
isort==5.13.2
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
kiwisolver==1.4.2
latexcodec==2.0.1
lightning==2.0.3
lightning-cloud==0.5.70
lightning-utilities==0.10.1
lit==16.0.6
littleutils==0.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matbench==0.6
matminer==0.7.4
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mkl==2023.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
monty==2022.4.26
mp-api==0.35.1
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.5
munkres==1.1.4
mypy==1.14.1
mypy-extensions==1.0.0
nbformat==5.7.0
nest-asyncio==1.5.5
networkx==3.1
numexpr==2.8.3
numpy==1.22.4
numpydoc==1.6.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
optuna==3.6.1
optuna-integration==3.6.0
ordered-set==4.1.0
outdated==0.2.2
packaging==24.2
palettable==3.3.3
pandas==1.4.4
parso==0.8.3
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.21.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.18.0
pluggy==1.3.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pybtex==0.24.0
pycairo==1.21.0
pycodestyle==2.12.1
pycparser==2.21
pydantic==1.10.9
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.2.7
pymatgen==2023.8.10
pymongo==4.6.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.4.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
rdkit-pypi==2022.9.5
readchar==4.1.0
regex==2023.6.3
reportlab==3.5.67
requests==2.28.1
rich==13.7.1
rlPyCairo @ file:///home/conda/feedstock_root/build_artifacts/rlpycairo_1687519531733/work
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
runs==1.2.2
s3transfer==0.10.2
safetensors==0.3.1
scikit-learn==1.0.1
scipy==1.7.3
sentry-sdk==1.9.0
setproctitle==1.2.2
shortuuid==1.0.9
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==4.0.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
spglib==2.2.0
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.39
stack-data==0.2.0
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tbb==2021.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.13.2
torch==2.0.1
torch_geometric==2.5.0
torchaudio==2.0.2
torchdata==0.7.1
torchmetrics==1.3.0.post0
torchvision==0.15.2
tornado==6.2
tqdm==4.64.1
traitlets==5.14.3
transformers==4.30.2
triton==2.0.0
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1680081134351/work
uncertainties==3.1.7
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1667239955025/work
urllib3==1.26.12
uvicorn==0.30.3
wandb==0.13.4
wcwidth==0.2.5
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.3
xmod==1.8.1
yarl==1.9.4
zipp==3.10.0
